requirements.txt
pytest~=7.1.3
webdriver_manager~=3.8.3
selenium~=4.5.0
rstr~=3.2.0
pytest-xdist~=2.5.0
allure-pytest~=2.11.1
Appium-Python-Client~=2.7.1